exosphere-runtimes/cloud-storage-runtime/pyproject.toml
[project]
name = "cloud-storage-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.1",
    "python-dotenv>=1.1.1",
]
